Add I2C/VL53L0X example
This is a minimal example that sets up I2C and uses it to talk to the
VL53L0X distance sensor. It reads a reference value from the sensor and
compares it to the expected value, to test out the I2C code.

This is a first step towards implementing an I2C API. The next step
would be to factor out the I2C code and move it from the example to the
library crate.